+++ jsonrpc-spring-boot-autoconfigure/src/test/java/com/limehee/jsonrpc/spring/boot/autoconfigure/JsonRpcMethodAccessInterceptorTest.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.node.IntNode;
 import com.limehee.jsonrpc.core.JsonRpcErrorCode;
 import com.limehee.jsonrpc.core.JsonRpcException;
 import com.limehee.jsonrpc.core.JsonRpcRequest;
+import com.limehee.jsonrpc.spring.boot.autoconfigure.support.JsonRpcMethodAccessInterceptor;
 import org.junit.jupiter.api.Test;
 import org.springframework.core.Ordered;